+++ inst/examples/getThis.r
@@ -1,0 +1,7 @@
+\dontrun{
+
+where <- new.env()  
+setThis(id = "x_1", value = TRUE, where = where)
+getThis(id = "x_1", where = where)
+
+}