+++ inst/examples/getValue.r
@@ -1,7 +1,0 @@
-\dontrun{
-
-where <- new.env()  
-setValue(id = "x_1", value = TRUE, where = where)
-getValue(id = "x_1", where = where)
-
-}