+++ tests/testthat/test-getThis-1.r
@@ -1,0 +1,11 @@
+context("getThis-1")
+test_that("getThis", {
+
+  ## Create example content //
+  where <- new.env()
+  setThis(id = "test", value = new.env(), where = where, 
+           binding_type = 2)
+  expect_is(getThis(id = "test", where = where), "environment")
+  
+  }
+)