+++ NEWS.md
@@ -1,3 +1,22 @@
+# CHANGES IN reactr VERSION 0.1.0.5
+
+## NEW FEATURES
+
+## BUG FIXES
+
+- fixed documentation in `R/setThis_bare.r`
+
+## MAJOR CHANGES
+
+- changed `setValue*()` to `setThis*()`
+- changed `getValue()` to `getThis()`
+
+## MINOR CHANGES
+
+## MISC
+
+-----
+
 # CHANGES IN reactr VERSION 0.1.0.4
 
 ## NEW FEATURES
@@ -22,7 +41,7 @@
 
 ## MAJOR CHANGES
 
-- added S3 bare function `setValue_bare`
+- added S3 bare function `setThis_bare`
 - added argument `mutual` and implemented functionality to set mutual bindings
 - added argument `where_watch` 
 - added argument `.hash_id` 
@@ -46,7 +65,7 @@
 ## MAJOR CHANGES
 
 - argument `envir` changed to `where` in order to make the interfaces for 
-  `setValue` and `getValue` more generic
+  `setThis` and `getThis` more generic
 - added dependencies for `classr` and `digest`
 - started to implement mutual bindings (not fully functional yet)